trying not to cry

+++ src/main/java/br/com/gabriel/entities/Entity.java
@@ -7,5 +7,5 @@ import java.io.Serializable;
  * @author gabri
  */
 public interface Entity extends Serializable {
-  public abstract Integer getId();
+  public abstract Long getId();
 }
